pyproject.toml
[project]
name = "gmail-classifier"
version = "0.1.0"
description = "AI Agent for classifying Gmail messages using LLMs."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.13",
    "langchain-google-vertexai>=0.1.1",
    "langchain-openai>=0.1.1",
    "langchain-anthropic>=0.1.4",
    "google-api-python-client>=2.122.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "pydantic>=2.6.4",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-cov>=5.0.0",
    "ruff>=0.3.4",
    "mypy>=1.9.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = [
    "googleapiclient.*",
    "google.auth.*",
    "google.oauth2.*"
]
ignore_missing_imports = true
